Sketch out a openslide loader for libdicom
This is a prototype WSI DICOM loader for openslide, using libdicom.

It can correctly load TILED_FULL dicom images and includes support for label and overview images. The 3D Histech demo DICOMs work well.

You give it the filename of one of the DICOMs in the pyramid you want to open, it then searches the enclosing directory and finds the other pyramid levels and any additional images with a matching slide id.

Missing features:

- sparse images
- metadata support
- other tile formats (it's only JPEG right now)
- ICC support (though I think this should perhaps be a different PR)
- I added libdicom to the ubuntu CI build, but I've not been able to get the windows build working yet
- tests
- enable fedora test builds
- openslide website updates (again, a separate PR I guess)

Signed-off-by: John Cupitt <jcupitt@gmail.com>